Balanced things a bit, and begun to work on a menu.
THIS DOES NOT BUILD :(

+++ Project 33/Assets/Scripts/Controllers/Bullet.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Bullet : MonoBehaviour {
     public int bulletSpeed = 50;
-    public int bulletDamage = 1;
+    public float bulletDamage = 1;
     public GameObject hitAnimation;
     Vector3 bulletVelocity;
